Reconfigure of main complete.
